refactor way settings are registered to allow reusability
