+++ includes/classes/Feature/VectorEmbeddings/VectorEmbeddings.php
@@ -69,11 +69,6 @@ class VectorEmbeddings extends Feature {
 			'elasticpress-labs'
 		);
 
-		// Set up settings page sub-module
-		if ( $this->is_active() ) {
-			$this->settings_page = new Settings();
-			$this->settings_page->setup();
-		}
 		parent::__construct();
 	}
